pratt: clean-up by pulling op-related into own mod
